+++ Anarchy-BepInEx/AnarchyMod.cs
@@ -91,7 +91,9 @@ namespace Anarchy
             updateSystem.UpdateBefore<RemoveOverridenSystem>(SystemUpdatePhase.ModificationEnd);
             updateSystem.UpdateAt<PreventCullingSystem>(SystemUpdatePhase.ToolUpdate);
 
-            updateSystem.UpdateBefore<ModifyNetCompositionDataSystem>(SystemUpdatePhase.Modification4);
+            updateSystem.UpdateBefore<ModifyNetCompositionDataSystem>(SystemUpdatePhase.Modification1);
+            updateSystem.UpdateBefore<ResetNetCompositionDataSystem>(SystemUpdatePhase.Modification5);
+            updateSystem.UpdateBefore<ModifyNetCompositionDataSystem>(SystemUpdatePhase.ModificationEnd);
             updateSystem.UpdateAfter<ResetNetCompositionDataSystem>(SystemUpdatePhase.ModificationEnd);
         }