+++ Anarchy-BepInEx/Systems/ModifyNetCompositionDataSystem.cs
@@ -98,6 +98,7 @@ namespace Anarchy.Systems
                 }
             }
 
+            m_ResetNetCompositionDataSystem.Enabled = true;
             entities.Dispose();
         }
     }