+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 * All bulldozing aspects of the mod have been forked over to a new mod called Better Bulldozer.
 * While using the net tool Anarchy will now let you violate the clearance of other networks.
 * Fixed accidentally adding the PreventOverride component to vehicles, cims, households, events, and buildings. This caused various serious issues. This version should remove all of the improperly added components and not add new ones to those types of entities.
-
+* Fixed plugin data
+* Updated Logo
+ 
 ## Patch V1.2.2
 * Added trees and plants back into query for PreventCullingSystem.